+++ api/src/main/java/org/pmiops/workbench/monitoring/GaugeRecorderService.java
@@ -3,11 +3,13 @@ package org.pmiops.workbench.monitoring;
 import com.google.common.collect.ImmutableMap;
 import java.util.List;
 import java.util.Map;
+import java.util.logging.Logger;
 import org.pmiops.workbench.monitoring.views.OpenCensusStatsViewInfo;
 import org.springframework.stereotype.Service;
 
 @Service
 public class GaugeRecorderService {
+  private static final Logger logger = Logger.getLogger(GaugeRecorderService.class.getName());
 
   private final List<GaugeDataCollector> gaugeDataCollectors;
   private final MonitoringService monitoringService;
@@ -25,6 +27,19 @@ public class GaugeRecorderService {
             .flatMap(m -> m.entrySet().stream())
             .collect(ImmutableMap.toImmutableMap(Map.Entry::getKey, Map.Entry::getValue));
     monitoringService.recordValues(result);
+    logValues(result); // temporary until dashboards are up and robust
     return result;
+  }
+
+  private void logValues(Map<OpenCensusStatsViewInfo, Number> viewToValue) {
+    StringBuilder textBuilder = new StringBuilder();
+    viewToValue.forEach(
+        (view, value) ->
+            textBuilder
+                .append(view.getStatsName().toString())
+                .append(" = ")
+                .append(value.toString())
+                .append("\n"));
+    logger.info(textBuilder.toString());
   }
 }
